+++ apps/src/aiTutor/views/teacherDashboard/InteractionsTable.tsx
@@ -17,6 +17,7 @@ import styleConstants from '@cdo/apps/styleConstants';
 import SafeMarkdown from '@cdo/apps/templates/SafeMarkdown';
 import {tableLayoutStyles as style} from '@cdo/apps/templates/tables/tableConstants';
 import color from '@cdo/apps/util/color';
+import i18n from '@cdo/locale';
 
 // TODO: Condense use of inline and imported styles
 import interactionsStyle from './interactions-table.module.scss';
@@ -260,10 +261,12 @@ const InteractionsTable: React.FC<InteractionsTableProps> = ({sectionId}) => {
           labelText="Filter by Status"
           name="filter-statuses"
           onChange={handleStatusFilterChange}
-          onClearAll={() => setSelectedStatuses([])}
           onSelectAll={() =>
             setSelectedStatuses(Object.values(AITutorInteractionStatus))
           }
+          onClearAll={() => setSelectedStatuses([])}
+          selectAllText={i18n.selectAll()}
+          clearAllText={i18n.clearAll()}
           size="s"
         />
         <CheckboxDropdown
@@ -274,10 +277,12 @@ const InteractionsTable: React.FC<InteractionsTableProps> = ({sectionId}) => {
           labelText="Filter by Student"
           name="filter-students"
           onChange={handleStudentFilterChange}
-          onClearAll={() => setSelectedUserIds([])}
           onSelectAll={() =>
             setSelectedUserIds(studentFilterOptions.map(option => option.value))
           }
+          onClearAll={() => setSelectedUserIds([])}
+          selectAllText={i18n.selectAll()}
+          clearAllText={i18n.clearAll()}
           size="s"
         />
         <SimpleDropdown