+++ apps/src/componentLibrary/tags/_Tag.tsx
@@ -1,29 +1,9 @@
 import React, {memo} from 'react';
-import {
-  OverlayTrigger,
-  Tooltip,
-  TooltipProps,
-  OverlayTriggerProps,
-} from 'react-bootstrap-2'; // TODO: Once we have [DSCO] Tooltip component, replace this import with it
 
 import FontAwesomeV6Icon from '@cdo/apps/componentLibrary/fontAwesomeV6Icon';
+import {WithTooltip} from '@cdo/apps/componentLibrary/tooltip';
 
 import moduleStyles from './tags.module.scss';
-
-// Allow the tooltips to display on focus so that the information
-// can be shown via keyboard
-const LabelOverlayTrigger: React.FC<OverlayTriggerProps> = props => (
-  <OverlayTrigger placement="top" trigger={['hover', 'focus']} {...props} />
-);
-
-// The arrowProps passed down in ReactBootstrap use styles that
-// conflict with the custom styles that we want, so they
-// are extracted out here.
-const LabelTooltip = React.forwardRef<HTMLDivElement, TooltipProps>(
-  ({arrowProps, ...props}, ref: React.Ref<HTMLDivElement>) => (
-    <Tooltip ref={ref} {...props} />
-  )
-);
 
 type TagIconProps = {
   iconName: string;
@@ -60,20 +40,15 @@ const Tag: React.FunctionComponent<TagProps> = ({
   icon,
 }) => {
   return (
-    <LabelOverlayTrigger
-      overlay={(props: TooltipProps) => (
-        <LabelTooltip
-          id={tooltipId}
-          className={moduleStyles.tagTooltip}
-          {...props}
-        >
-          {tooltipContent}
-        </LabelTooltip>
-      )}
+    <WithTooltip
+      tooltipProps={{
+        direction: 'onTop',
+        text: tooltipContent,
+        tooltipId,
+      }}
     >
       <div
         tabIndex={0} // eslint-disable-line jsx-a11y/no-noninteractive-tabindex
-        role="tooltip"
         aria-describedby={tooltipId}
         className={moduleStyles.tag}
       >
@@ -81,7 +56,7 @@ const Tag: React.FunctionComponent<TagProps> = ({
         <span>{label}</span>
         {icon && icon.placement === 'right' && <TagIcon {...icon} />}
       </div>
-    </LabelOverlayTrigger>
+    </WithTooltip>
   );
 };