+++ apps/src/templates/SchoolDataInputs.jsx
@@ -26,6 +26,8 @@ const SEARCH_DEFAULTS = [
 
 const COUNTRIES_US_FIRST = getCountriesUsFirst();
 
+export const SCHOOL_INFO_ID = 'school_info';
+
 export default function SchoolDataInputs({
   schoolId,
   country,
@@ -37,6 +39,7 @@ export default function SchoolDataInputs({
   setSchoolName,
   setSchoolZip,
   usIp,
+  containerClassName,
   includeHeaders = true,
   markFieldsAsRequired = false,
   fieldNames = {
@@ -89,12 +92,16 @@ export default function SchoolDataInputs({
 
   const labelClassName = schoolZipIsValid ? '' : style.disabledLabel;
 
-  const computedStyleClass = classNames(style.schoolAssociationWrapper, {
-    [style.requiredLabel]: markFieldsAsRequired,
-  });
+  const computedStyleClass = classNames(
+    style.schoolAssociationWrapper,
+    {
+      [style.requiredLabel]: markFieldsAsRequired,
+    },
+    containerClassName
+  );
 
   return (
-    <div className={computedStyleClass}>
+    <div id={SCHOOL_INFO_ID} className={computedStyleClass}>
       {includeHeaders && (
         <div className={style.headerContainer}>
           <Heading2>{i18n.censusHeading()}</Heading2>
@@ -212,6 +219,7 @@ SchoolDataInputs.propTypes = {
   includeHeaders: PropTypes.bool,
   markFieldsAsRequired: PropTypes.bool,
   fieldNames: PropTypes.object,
+  containerClassName: PropTypes.string,
   schoolId: PropTypes.string.isRequired,
   country: PropTypes.string.isRequired,
   schoolName: PropTypes.string.isRequired,