+++ apps/src/componentLibrary/popover/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.0](https://github.com/code-dot-org/code-dot-org/pull/63073)
+
+- updated `Popover` props to support native HTML Div element attributes
+
 ## [0.2.1](https://github.com/code-dot-org/code-dot-org/pull/62920)
 
 - updated color variables to use `primitiveColors.css`