+++ apps/src/componentLibrary/tags/CHANGELOG.md
@@ -2,8 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.0](https://github.com/code-dot-org/code-dot-org/pull/62996)
+
+- refactored `Tags` tp use DSCO `Tooltip` instead of react-bootstrap `Tooltip`
+- removed unneeded styles
+- updated color variables to use `primitiveColors.css`
+- rewritten `TagsTest` in typescript
+- refactored/optimized and added more tests to `TagsTest.tsx`
+
 ## [0.2.0](https://github.com/code-dot-org/code-dot-org/pull/58130)
-* Added `tabIndex={0}` to ensure tooltip shows up during tab navigation. See PR for more discussion.
+
+- Added `tabIndex={0}` to ensure tooltip shows up during tab navigation. See PR for more discussion.
 
 ## [0.1.0](https://github.com/code-dot-org/code-dot-org/pull/55420)
-* Initial release
+
+- Initial release